[Refactor] change spotify service dir
