+++ src/services/playlist/spotify/index.js
@@ -1,0 +1,7 @@
+const playList = require("./playList");
+const track = require("./track");
+
+module.exports = {
+  playList,
+  track,
+};
