+++ src/services/playlist/spotify/track.js
@@ -1,40 +1,8 @@
 const axios = require("axios");
-const { cacheService } = require("../database");
 
-const searchList = async (token) => {
-  try {
-    const params = {
-      limit: 50,
-    };
-    const options = {
-      method: "GET",
-      url: "https://api.spotify.com/v1/me/playlists",
-      headers: {
-        authorization: `Bearer ${token}`,
-      },
-      params,
-    };
+const { cacheService } = require("../../database");
 
-    const playLists = [];
-    do {
-      const response = await axios(options);
-      const { data } = response;
-
-      data.items.forEach((item) => {
-        playLists.push(parsePlayList(item));
-      });
-
-      options.url = data.next;
-    } while (options.url);
-
-    return { playLists };
-  } catch (error) {
-    console.error(error);
-    throw new Error(error);
-  }
-};
-
-const getTrack = async (id, token) => {
+const get = async (id, token) => {
   try {
     const options = {
       method: "GET",
@@ -68,18 +36,7 @@ const getTrack = async (id, token) => {
   }
 };
 
-const parsePlayList = (playList) => {
-  return {
-    id: playList.id,
-    title: playList.name,
-    thumbnail: playList.images[0],
-    description: playList.description,
-    owner: {
-      name: playList.owner.display_name,
-      id: playList.owner.id,
-    },
-  };
-};
+module.exports = { get };
 
 const parseTrackItem = (track) => {
   const artists = [];
@@ -102,5 +59,3 @@ const parseTrackItem = (track) => {
     thumbnail: track.album.images[0],
   };
 };
-
-module.exports = { searchList, getTrack };
